Fix issue with types
