+++ .changeset/itchy-timers-fetch.md
@@ -1,0 +1,5 @@
+---
+"solid-echarts": patch
+---
+
+Make types less strict
