+++ src/SolidEChartsCore.tsx
