+++ src/index.ts
@@ -1,0 +1,22 @@
+import SolidECharts from "solid-echarts";
+import SolidEChartsCore from "./SolidEChartsCore";
+import {
+  SolidEChartsProps,
+  EChartsSetOptionOpts,
+  EChartsTheme,
+  EChartsInitOpts,
+  EChartsType,
+  EChartsOption,
+  EChartsInit,
+} from "./types";
+
+export type {
+  SolidEChartsProps,
+  EChartsSetOptionOpts,
+  EChartsTheme,
+  EChartsInitOpts,
+  EChartsType,
+  EChartsOption,
+  EChartsInit,
+};
+export { SolidECharts as default, SolidEChartsCore };
